+++ app/src/main/java/woowacourse/movie/BackKeyActionBarActivity.kt
@@ -1,0 +1,48 @@
+package woowacourse.movie
+
+import android.content.Intent
+import android.os.Build
+import android.os.Bundle
+import android.view.MenuItem
+import androidx.appcompat.app.AppCompatActivity
+import java.io.Serializable
+
+abstract class BackKeyActionBarActivity : AppCompatActivity() {
+
+    @Deprecated("deprecated", ReplaceWith("onCreateView"), DeprecationLevel.WARNING)
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        supportActionBar?.setDisplayHomeAsUpEnabled(true)
+        onCreateView(savedInstanceState)
+    }
+
+    abstract fun onCreateView(savedInstanceState: Bundle?)
+
+    @Suppress("DEPRECATION")
+    inline fun <reified T : Serializable> Intent.customGetSerializable(key: String): T? {
+        return if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.TIRAMISU) {
+            getSerializableExtra(key, T::class.java)
+        } else {
+            getSerializableExtra(key) as? T
+        }
+    }
+
+    @Suppress("DEPRECATION")
+    inline fun <reified T : Serializable> Bundle.customGetSerializable(key: String): T? {
+        return if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.TIRAMISU) {
+            getSerializable(key, T::class.java)
+        } else {
+            getSerializable(key) as? T
+        }
+    }
+
+    override fun onOptionsItemSelected(item: MenuItem): Boolean {
+        return when (item.itemId) {
+            android.R.id.home -> {
+                finish()
+                true
+            }
+            else -> super.onOptionsItemSelected(item)
+        }
+    }
+}
